Add delete and update functions

+++ app/controllers/artifact-controller.js
@@ -24,8 +24,8 @@
 // circular references between controllers.
 module.exports = {
   createArtifact,
-  deleteArtifact
-  //updateArtifact
+  deleteArtifact,
+  updateArtifact
 
 };
 
@@ -43,21 +43,14 @@ const utils = M.require('lib.utils');
 /**
  * @description This function creates an Artifact.
  *
- * @param {User} reqUser  The user object of the requesting user.
- * @param {Object} artifactMetaData  The JSON object containing the Artifact data
- * @param {Binary} artifactBlob  The binary data to store.
+ * @param {User} reqUser - The user object of the requesting user.
+ * @param {Project} proj - The project object of the artifact.
+ * @param {Organization} org - The Org object of the artifact.
+ * @param {Object} artifactMetaData - The JSON object containing the Artifact data
+ * @param {Binary} artifactBlob - The binary data to store.
  *
  * @return {Promise} resolve - new Artifact
  *                   reject - error
- *
- * @example
- * createArtifact({User}, { id: 'elementID', project: { id: 'projID', org: {id: 'orgID' }}})
- * .then(function(Artifact) {
- *   // Do something with the newly created Artifact
- * })
- * .catch(function(error) {
- *   M.log.error(error);
- * });
  */
 function createArtifact(reqUser, org, proj, artifactMetaData, artifactBlob) {
   return new Promise((resolve, reject) => {
@@ -74,36 +67,165 @@ function createArtifact(reqUser, org, proj, artifactMetaData, artifactBlob) {
       return reject(new M.CustomError(error.message, statusCode, 'warn'));
     }
 
-    // Generate hash
-    let hashedName = mbeeCrypto.sha256Hash(artifactBlob);
-    let fileDest = path.join(M.artifactPath, hashedName);
+    // Create the full artifact id
+    artifactFullId = utils.createUID(org.id, proj.id, artifactMetaData.id);
 
-    // Check if hash folder already exist
-    fs.exists(fileDest, (exists) => {
-      // Check directory NOT exist
-      if (!exists) {
-        M.log.info('Adding new artifact...');
-
-        // Add artifact to storage
-        addArtifactOS(hashedName, artifactMetaData.filename, artifactBlob);
+    // Find artifact in database
+    findArtifact(reqUser, artifactFullId)
+    .then((_artifact) => {
+      // Error Check: ensure no matching artifact already exist
+      if (typeof _artifact !== 'undefined' && _artifact) {
+        // One or more users exists, reject
+        return reject(new M.CustomError('Artifact(s) with matching Id '
+          + ` already exist: [${_artifact.id}].`, 403, 'warn'));
       }
-      else{
-        // Folder exist, no need to duplicate to file system
-        M.log.info('Existing Artifact...');
+      // Generate hash
+      let hashedName = mbeeCrypto.sha256Hash(artifactBlob);
+      let fileDest = path.join(M.artifactPath, hashedName);
+
+      // Check if hash folder already exist in filesystem
+      fs.exists(fileDest, (exists) => {
+        // Check directory NOT exist
+        if (!exists) {
+          M.log.info('Adding new artifact...');
+
+          // Add artifact to storage
+          addArtifactOS(hashedName, artifactMetaData.filename, artifactBlob);
+        }
+        else{
+          // Folder exist, no need to duplicate to file system
+          M.log.info('Existing Artifact...');
+        }
+      });
+
+      // Define new hash history data
+      const historyData = {
+        hash: hashedName,
+        user: reqUser
       }
+
+      // Create the new Artifact
+      const artifact = new Artifact({
+        id: utils.createUID(org.id, proj.id, artifactMetaData.id),
+        filename: artifactMetaData.filename,
+        contentType: path.extname(artifactMetaData.filename),
+        history: historyData,
+        project: proj,
+        organization:org
+
+      });
+
+      // Save user object to the database
+      return artifact.save();
+    })
+    .then((_artifact) => resolve(_artifact))
+    .catch((error) => {
+      M.log.error(error);
+      // Error occurred, reject it
+      return reject(error);
     });
+  });
+}
 
-    // Create the new Artifact
-    const artifact = new Artifact({
-      id: utils.createUID(org.id, proj.id, artifactMetaData.id),
-      filename: artifactMetaData.filename,
-      hash: hashedName,
-      contentType: 'png'
+/**
+ * @description This function updates an existing Artifact.
+ *
+ * @param {User} reqUser - The user object of the requesting user.
+ * @param {Project} proj - The project object of the artifact.
+ * @param {Organization} org - The Org object of the artifact.
+ * @param {Object} artifactToUpdate - JSON object containing updated Artifact data
+ * @param {Binary} artifactBlob - The binary data to store.
+ *
+ * @return {Promise} resolve - new Artifact
+ *                   reject - error
+ */
+function updateArtifact(reqUser, org, proj, artifactToUpdate, artifactBlob) {
+  return new Promise((resolve, reject) => {
+    // Error Check: ensure input parameters are valid
+    try {
+      assert.ok(reqUser.admin, 'User does not have permissions.');
+    }
+    catch (error) {
+      let statusCode = 400;
+      // Return a 403 if request is permissions related
+      if (error.message.includes('permissions')) {
+        statusCode = 403;
+      }
+      return reject(new M.CustomError(error.message, statusCode, 'warn'));
+    }
 
-    });
+    // Create the full artifact id
+    artifactFullId = utils.createUID(org.id, proj.id, artifactToUpdate.id);
 
-    // Save user object to the database
-    artifact.save()
+    // Find artifact in database
+    findArtifact(reqUser, artifactFullId)
+    .then((_artifact) => {
+      // Error Check: artifact was NOT found
+      if (!_artifact) {
+        // Artifact not found, reject with error
+        return reject(new M.CustomError('Artifact not found', 404, 'warn'));
+      }
+      
+      // Generate hash
+      let hashedName = mbeeCrypto.sha256Hash(artifactBlob);
+      let fileDest = path.join(M.artifactPath, hashedName);
+
+      // Check if hash folder already exist in filesystem
+      fs.exists(fileDest, (exists) => {
+        // Check directory NOT exist
+        if (!exists) {
+          M.log.info('Adding new artifact...');
+
+          // Add artifact to storage
+          addArtifactOS(hashedName, artifactToUpdate.filename, artifactBlob);
+        }
+        else{
+          // Folder exist, no need to duplicate to file system
+          M.log.info('Existing Artifact...');
+        }
+      });
+
+      // Remove 'id' list
+      // Note: id is immutable and cannot be changed
+      delete artifactToUpdate['id'];
+
+      // Define and get a list of artifact keys to update
+      const artifactUpdateFields = Object.keys(artifactToUpdate);
+
+      // Get list of parameters which can be updated from model
+      const validUpdateFields = _artifact.getValidUpdateFields();
+
+      // Loop through all updateable fields
+      for (let i = 0; i < artifactUpdateFields.length; i++) {
+        // Error Check: Check if field can be updated
+        if (!validUpdateFields.includes(artifactUpdateFields[i])) {
+          // field cannot be updated, reject error
+          return reject(new M.CustomError(
+            `Artifact property [${artifactUpdateFields[i]}] cannot be changed.`, 403, 'warn'
+          ));
+        }
+        // Sanitize field and update field in user object
+        _artifact[artifactUpdateFields[i]] =
+          sani.sanitize(artifactToUpdate[artifactUpdateFields[i]]);
+      }
+
+      // Get history length
+      const historyIndex = _artifact.history.length-1;
+      // Check hash has changed
+      if (hashedName !== _artifact.history[historyIndex].hash)
+      {
+        // Define new hash history data
+        const historyData = {
+          hash: hashedName,
+          user: reqUser
+        };
+        // Hash changed, add to history
+        _artifact.history.push(historyData);
+      }
+
+      // Save user object to the database
+      return _artifact.save();
+    })
     .then((_artifact) => resolve(_artifact))
     .catch((error) => {
       M.log.error(error);
@@ -117,19 +239,13 @@ function createArtifact(reqUser, org, proj, artifactMetaData, artifactBlob) {
  * @description This function deletes an Artifact.
  *
  * @param {User} reqUser  The user object of the requesting user.
- * @param {Object} element  The JSON object containing the Artifact data
+ * @param {Project} proj - The project object of the artifact.
+ * @param {Organization} org - The Org object of the artifact.
+ * @param {String} artifactId - The Artifact Id
  *
  * @return {Promise} resolve - new Artifact
  *                   reject - error
  *
- * @example
- * createArtifact({User}, { id: 'elementID', project: { id: 'projID', org: {id: 'orgID' }}})
- * .then(function(Artifact) {
- *   // Do something with the newly created Artifact
- * })
- * .catch(function(error) {
- *   M.log.error(error);
- * });
  */
 function deleteArtifact(reqUser, org, proj, artifactId) {
   return new Promise((resolve, reject) => {
@@ -156,27 +272,37 @@ function deleteArtifact(reqUser, org, proj, artifactId) {
 
     // Find to be deleted artifact
     findArtifact(reqUser, artifactFullId)
-    .then((artifact) => {
-      artifactToDelete = artifact;
-      hashedName = artifact.hash;
-      filename = artifact.filename;
+    .then((artifactToDelete) => {
+      // Error Check: artifact is NOT found
+      if (!artifactToDelete) {
+        // No artifact exists, reject
+        return reject(new M.CustomError('Artifact not found.', 404, 'warn'));
+      }
+      // Define and get artifact history
+      const artifactHistory = artifactToDelete.history;
+      const filename = artifactToDelete.filename;
 
+      // Loop through object history
+      for (let i = 0; i < artifactHistory.length; i++) {
+        // Find Artifact in filesystem
+        // Check no db record linked to this artifact
+        Artifact.find({ 'history.hash': artifactHistory[i].hash })
+        .then((remainingArtifacts) => {
+          // Check last record with this hash
+          if (remainingArtifacts.length === 1) {
+            // Last hash record, remove artifact to storage
+            removeArtifactOS(artifactHistory[i].hash, filename);
+          }
+        })
+        .catch((error) => reject(error));
+      }
+      // Remove artifact from database
       return artifactToDelete.remove();
     })
-    .then((_artifact) => {
-      // Find Artifact
-      // Check no db record linked to this artifact
-      Artifact.find({ hash: hashedName })
-      .then((remainingArtifacts) => {
-        // Check no record with this hash
-        if (remainingArtifacts.length === 0) {
-          // Last hash record, remove artifact to storage
-          removeArtifactOS(hashedName, filename);
-        }
-      })
-      .catch((error) => reject(error));
-
-      resolve(_artifact);
+    .then((_deletedArtifact) => {
+      // Return the shorten artifact ID
+      const shortenArtifactId = _deletedArtifact.id.split(':')[2];
+      resolve(shortenArtifactId);
     })
     .catch((error) => {
       M.log.error(error);
@@ -194,7 +320,8 @@ function deleteArtifact(reqUser, org, proj, artifactId) {
  *                delimited by colons Example: 'testorg00:project00:artifact00'
  * @param {Boolean} softDeleted - A boolean value indicating whether to soft delete.
  *
- * @returns {Promise} - resolve(artifact)/ reject(error)
+ * @return {Promise} resolve - new Artifact
+ *                   reject - error
  * */
 function findArtifact(reqUser, artifactFullId, softDeleted = false) {
   return new Promise((resolve, reject) => {
@@ -221,13 +348,7 @@ function findArtifact(reqUser, artifactFullId, softDeleted = false) {
     // Find Artifact
     Artifact.find(searchParams)
     .then((artifact) => {
-      // Error Check: ensure at least one artifact was found
-      if (artifact.length === 0) {
-        // No artifact found, reject error
-        return reject(new M.CustomError('Artifact not found.', 404, 'warn'));
-      }
-
-      // All checks passed, resolve artifact
+      // Return resolve artifact
       return resolve(artifact[0]);
     })
     .catch((error) => reject(new M.CustomError(error.message), 500, 'warn'));
@@ -243,8 +364,8 @@ function findArtifact(reqUser, artifactFullId, softDeleted = false) {
 function addArtifactOS(hashName, filename, artifactBlob) {
   // Create folder directory
   // Note: Folder name is the first 2 characters from the generated hash
-  let folderPath = path.join(M.artifactPath, hashName.substring(0,2));
-  let artifactPath = path.join(folderPath, hashName);
+  const folderPath = path.join(M.artifactPath, hashName.substring(0,2));
+  const artifactPath = path.join(folderPath, hashName);
 
   // Check file exist
   fs.exists(folderPath, (exists) => {
@@ -280,8 +401,8 @@ function addArtifactOS(hashName, filename, artifactBlob) {
 function removeArtifactOS(hashName) {
   // Create folder directory
   // Note: Folder name is the first 2 characters from the generated hash
-  let folderPath = path.join(M.artifactPath, hashName.substring(0,2));
-  let artifactPath = path.join(folderPath, hashName);
+  const folderPath = path.join(M.artifactPath, hashName.substring(0,2));
+  const artifactPath = path.join(folderPath, hashName);
 
   // Remove the artifact file
   fs.unlink(artifactPath, (error) => {
@@ -291,12 +412,22 @@ function removeArtifactOS(hashName) {
     }
   });
 
-  // Remove the directory
-  fs.rmdir(folderPath, (error) => {
-    // Check directory NOT exist
-    if (error) {
-      throw new M.CustomError(error.message, 500, 'warn');
+  // Check if directory is empty
+  fs.readdir(folderPath, function(err, files) {
+    if (err) {
+      // some sort of error
+    } else {
+      // Check if empty directory
+      if (!files.length) {
+        // directory appears to be empty
+        // Remove the directory
+        fs.rmdir(folderPath, (error) => {
+          // Check directory NOT exist
+          if (error) {
+            throw new M.CustomError(error.message, 500, 'warn');
+          }
+        });
+      }
     }
   });
 }
-
